### 2017-08-30 #385 @whedon
[PRE REVIEW]: hei: Calculate Healthy Eating Index (HEI) Scores
**Submitting author:** @timfolsom (<a href="http://orcid.org/0000-0002-8647-5810">Timothy Folsom</a>)
**Repository:** <a href="https://github.com/timfolsom/hei" target ="_blank">https://github.com/timfolsom/hei</a>
**Version:** v0.1.0
