### 2017-09-24 #417 @whedon
[REVIEW]: hei: Calculate Healthy Eating Index (HEI) Scores
**Submitting author:** @timfolsom (<a href="http://orcid.org/0000-0002-8647-5810">Timothy Folsom</a>)
**Repository:** <a href="https://github.com/timfolsom/hei" target ="_blank">https://github.com/timfolsom/hei</a>
**Version:** v0.1.0
**Archive:** <a href="http://dx.doi.org/10.6084/m9.figshare.5469208" target="_blank">10.6084/m9.figshare.5469208</a>
