### 2017-09-25 vpnagraj/hei#9 @maelle
Add a description line to the repository
so that one sees something more informative than "No description, website, or topics provided." when landing on the page. :-)

This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/417).
