### 2017-09-25 vpnagraj/hei#10 @maelle
Installation instructions
Any reason why the current installation instructions mention Github only and not the easier `install.packages`?

A possible formulation (including the link to `devtools`) would be:

Install the package with:

```r
install.packages("hei")
```

Or install the development version using [devtools](https://github.com/hadley/devtools) with:

```r
library("devtools")
install_github("timfolsom/hei")
```
This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/417).
