### 2017-09-25 vpnagraj/hei#11 @maelle
Small typo in hei.R
Uncovered via `devtools::spell_check()`
"this can be overidden with the \code{verbose} parameter", overidden should be "overridden".

This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/417).
