### 2017-09-25 vpnagraj/hei#12 @maelle
Goodpractice output 
I've run `goodpractice::gp()` on your package, here is the output. I feel very strongly about the URL and BugReports fields in DESCRIPTION (since you have the Github repo and issue tracker, it's better to have them in DESCRIPTION so that users who find the package from CRAN can find them), but include everything for information.

```r
It is good practice to

  ✖ omit "Date" in DESCRIPTION. It is not required and it gets invalid quite often. A
    build date will be added to the package when you perform `R CMD build` on it.
  ✖ add a "URL" field to DESCRIPTION. It helps users find information about your package
    online. If your package does not have a homepage, add an URL to GitHub, or the CRAN
    package package page.
  ✖ add a "BugReports" field to DESCRIPTION, and point it to a bug tracker. Many online
    code hosting services provide bug trackers for free, https://github.com,
    https://gitlab.com, etc.
  ✖ use '<-' for assignment instead of '='. '<-' is the standard, and R users and
    developers are used it and it is easier to read your code for them if you use '<-'.

    R\combo.R:35:18
    R\hei.R:152:19
    R\hei.R:156:13

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor
    windows that are about 80 characters wide. Try make your lines shorter than 80
    characters

    R\combo.R:31:1
    R\combo.R:32:1
    R\combo.R:36:1
    R\get_diet.R:76:1
    R\get_diet.R:92:1
    ... and 2 more lines
```

This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/417).
