### 2017-09-25 vpnagraj/hei#13 @maelle
Release CRAN versions
I've just noted that there is no release in the Github repo although the package is on CRAN. In the future, for archiving you might consider releasing each CRAN version and linking it to NEWS.md.

This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/417).
