### 2017-09-25 vpnagraj/hei#14 @maelle
Functional claim
I see the test "HEI score is same as original validation data" which is great. I even think it should be mentioned explicitely in paper.md (i.e. that it has been validated that the R code produces the same output as the SAS original script).

This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/417).
