### 2017-09-25 vpnagraj/hei#15 @maelle
Code of conduct
On top of having the contributing guidelines you could add a code of conduct (e.g. via `devtools::use_code_of_conduct()`) to show this repo is a friendly place. :-)

This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/417).
