### 2017-09-25 vpnagraj/hei#16 @maelle
Create a pkgdown website?
Since you have written two useful vignettes, why not create a [`pkgdown`](https://github.com/hadley/pkgdown) website to make the whole documentation more easily browsable? This is obviously only a suggestion (like many of the issues I've opened, I'll make a summary) but I've had an excellent experience with `pkgdown`. If you do create one, add its URL to the Github repository URL field.

This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/417).
