### 2017-09-25 vpnagraj/hei#17 @maelle
Add a link to the dplyr website
When mentioning it in the getting started vignette.

This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/417).
