### 2017-09-25 vpnagraj/hei#16 @maelle
Added benefit if you create a `pkgdown` website, when writing "An example of how to do this is provided in a second vignette." in the first vignette you could have a direct link to that second vignette.
