### 2017-09-25 vpnagraj/hei#18 @maelle
Small bug in the Analysis.Rmd vignette
It lacks the header

```
output: rmarkdown::html_vignette
vignette: >
  %\VignetteEngine{knitr::knitr}
  %\VignetteIndexEntry{hei Vignette}
  %\usepackage[UTF-8]{inputenc}
```

This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/417).
