### 2017-09-25 vpnagraj/hei#18 @maelle
I noticed this because it wasn't built when using `devtools::build_vignettes`; instead, the function mentioned the missing header.
