### 2017-09-25 vpnagraj/hei#19 @maelle
Make target audience a bit more explicit?
In paper.md, it might be good to say which research fields use the score, even it's sort of implicit at the moment.

This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/417).
