### 2017-09-25 vpnagraj/hei#20 @maelle
Add a sentence more about the food security variable in the Analysis.Rmd vignette
or a link to a page where someone can learn what it means.

This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/417).
