### 2017-09-25 vpnagraj/hei#21 @maelle
Another small typo in hei.R
Boolean should be capitalized.

This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/417).
