### 2017-09-25 vpnagraj/hei#22 @maelle
Make "must use valid year choice" user-friendlier
In the functions where the error message "must use valid year choice" can appear, it might be user-friendlier to replace it with "must use valid year choice, see the function documentation to know which ones are valid".

This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/417).
