### 2017-09-25 vpnagraj/hei#23 @maelle
Have you used the package in your own work?
If so, cite the corresponding paper(s) in paper.md

This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/417).
