### 2017-09-25 #417 @maelle
I'm done with the review, nice package! You'll see I've opened many issues but nearly everything in the review list is crossed.

* I think the installation instructions could be made a bit clearer, see [this issue](https://github.com/timfolsom/hei/issues/10)

* I've made some suggestions for the documentation https://github.com/timfolsom/hei/issues/9, https://github.com/timfolsom/hei/issues/11, https://github.com/timfolsom/hei/issues/16, https://github.com/timfolsom/hei/issues/17, https://github.com/timfolsom/hei/issues/20, https://github.com/timfolsom/hei/issues/21.

* I've also made some suggestions regarding the paper. https://github.com/timfolsom/hei/issues/14, https://github.com/timfolsom/hei/issues/19, https://github.com/timfolsom/hei/issues/23.

* I found a (really easy to fix) bug in a vignette https://github.com/timfolsom/hei/issues/18 so this one is an important issue.

* Some other issues are really more suggestions! https://github.com/timfolsom/hei/issues/12, https://github.com/timfolsom/hei/issues/13, https://github.com/timfolsom/hei/issues/15, https://github.com/timfolsom/hei/issues/16, https://github.com/timfolsom/hei/issues/22.

Happy to discuss further is something is unclear.
