### 2017-09-25 #417 @leeper
@maelle Thank you so much for your thorough and prompt review!

@timfolsom Can you please address the issues Maëlle has raised and ping me when you're done?
