### 2017-09-25 5a1ca9e
fixed typos in hei.R
