+++ man/hei.Rd
@@ -15,7 +15,7 @@ hei(fped, diet, demograph, agethresh = 2, verbose = FALSE)
 
 \item{agethresh}{numeric threshold for age in years of survey participants to be included; any individual less than the value specified will be excluded; defaults to 2}
 
-\item{verbose}{boolean indicating whether or not all columns from processed data should be output; default is \code{FALSE} designating only the following are included in the returned \code{data.frame}:
+\item{verbose}{Boolean indicating whether or not all columns from processed data should be output; default is \code{FALSE} designating only the following are included in the returned \code{data.frame}:
 \itemize{
 \item SEQN: Respondent sequence number
 \item RIDAGEYR: Best age in years of the sample person at time of HH screening. Individuals 85 and over are topcoded at 85 years of age
@@ -23,7 +23,7 @@ hei(fped, diet, demograph, agethresh = 2, verbose = FALSE)
 }}
 }
 \value{
-Object of class \code{data.frame}; defaults to only include columns for respondent identifier, age, and overall HEI score of each individual; this can be overidden with the \code{verbose} parameter to output all columns of the input data sets as well as 33 columns of calculated data related to HEI scoring and, significantly, a 70th column containing the total HEI score for each participant.
+Object of class \code{data.frame}; defaults to only include columns for respondent identifier, age, and overall HEI score of each individual; this can be overridden with the \code{verbose} parameter to output all columns of the input data sets as well as 33 columns of calculated data related to HEI scoring and, significantly, a 70th column containing the total HEI score for each participant.
 }
 \description{
 \code{hei} calculates a Health Eating Index (HEI) score for individuals in the National Health and Nutrition Examination Survey (NHANES) studies based on input dietary, demographic and Food Pattern Equivalent Database (FPED) data.
