### 2017-09-25 64a6a81
added link to dplyr in getting started vignette
