+++ vignettes/GettingStarted.Rmd
@@ -21,7 +21,7 @@ Note: this workflow assumes the National Health and Nutrition Examination Survey
 
 The Healthy Eating Index is a dietary metric that gauges adherence to the Dietary Guidelines for Americans. The USDA uses the index primarily to keep tabs on the diet quality of the US population. By design, the HEI is geared toward use with in-depth dietary surveys such as the National Health and Nutrition Examination Survey (NHANES) conducted biannually by the CDC. The USDA offers guidance to researchers who would be interested in using the HEI in their analyse s of dietary trends and other health factors. Documentation on the National Institute of Health's official website walks through the steps of computing HEI scores from NHANES data with SAS. This package draws inspiration from the associated SAS code. Its intended use is as a tool for R users to perform these kinds of analyses and is best understood as a companion to NHANES. Note: an R package called nhanesA exists, is useful for retrieving NHANES data sets, and is required by this package.
 
-To get started, first we load the hei package and the dplyr package. The latter, in case you are unfamiliar, provides access to a number of useful functions that, while not necessary, greatly simplify some of the data manipulation steps and as such is recommended and will be used throughout these examples. Some of the functions included are filter(), group_by(), and select(), as well as the ``` %>% ``` "pipe."
+To get started, first we load the hei package and the [dplyr](http://dplyr.tidyverse.org/) package. The latter, in case you are unfamiliar, provides access to a number of useful functions that, while not necessary, greatly simplify some of the data manipulation steps and as such is recommended and will be used throughout these examples. Some of the functions included are filter(), group_by(), and select(), as well as the ``` %>% ``` "pipe."
 
 ```{r, echo=TRUE}
 library(hei)
