### 2017-09-25 596ca0a
updates to installation instructions in README
