### 2017-09-25 692c7ba
fixed typo in README
