### 2017-09-25 9604245
fixed mispelling of install.packages() in README
