### 2017-09-25 804f655
added code of conduct via devtools::use_code_of_conduct()
