### 2017-09-25 b018bd9
removed date from DESCRIPTION
