+++ DESCRIPTION
@@ -1,7 +1,6 @@
 Package: hei
 Type: Package
 Title: Calculate Healthy Eating Index (HEI) Scores
-Date: 2017-08-05
 Version: 0.1.0
 Authors@R: c(person("VP", "Nagraj", email = "vpnagraj@virginia.edu", role =
     c("aut")), person("Tim", "Folsom", email = "timothyjfolsom@gmail.com", role =
