### 2017-09-25 075b664
DESCRIPTION now has BugReports and URL fields
