+++ DESCRIPTION
@@ -21,3 +21,5 @@ Suggests: knitr,
     dplyr,
     ggplot2
 VignetteBuilder: knitr
+URL: https://cran.r-project.org/package=hei
+BugReports: https://github.com/timfolsom/hei/issues
