### 2017-09-25 70454ed
breaking up long lines in get_diet.R paste0 statement
