### 2017-09-25 vpnagraj/hei#12 @vpnagraj
@maelle thanks so much for the heads up on `goodpractice::gp()`:

1. added URL & BugReports field 075b664
2. dealt with some long lines 70454ed
3. removed date field from DESCRIPTION b018bd9
4. only using `<-` now 07c29ec
