### 2017-09-25 702871c
better error messages for day and year choices handling
