### 2017-09-25 vpnagraj/hei#22 @vpnagraj
thanks @maelle that's a really good point!

including more informative error handling via 702871c
