### 2017-09-25 ea5ff7c
added link to current CRAN version in NEWS.md
