+++ NEWS.md
@@ -1,5 +1,5 @@
 # hei
 
-## 0.1
+## 0.1.0
 
-- initial release
+- initial release on CRAN: <https://CRAN.R-project.org/package=hei>
