### 2017-09-25 vpnagraj/hei#13 @vpnagraj
thanks @maelle ... now linking to CRAN version 0.1.0 of the package in the NEWS.md file ea5ff7c
