### 2017-09-25 8417706
added links regarding the food security data in the Analysis.Rmd vignette
