### 2017-09-25 4769b93
updates to paper including specifics about fields of interest
