+++ paper/paper.md
@@ -12,6 +12,6 @@ bibliography: paper.bib
 ---
 # Summary 
 
-The Health Eating Index (HEI) is a dietary metric originally designed by the United States Department of Agriculture (USDA) and National Cancer Institue (NCI) to gauge adherence to the United States Dietary Guidelines[@hei]. The NCI distributes SAS macros for calculating individual HEI scores. **hei** is a package that implements the HEI scoring algorithms in R[@R], allowing users without SAS licenses to perform these calculations for individuals across several years of National Health and Nutrition Examination Survey (NHANES) studies. Internally, **hei** depends on the **nhanesA** package[@nhanesA] to retrieve necessary data for scoring. Additional demographic and behavioral NHANES datasets are publicly available and can be combined with the HEI scores for corresponding individuals to serve a variety of research inquiries regarding dietary habits[@social_support][@tinnitus][@age]. 
+The Health Eating Index (HEI) is a dietary metric originally designed by the United States Department of Agriculture (USDA) and National Cancer Institute (NCI) to gauge adherence to the United States Dietary Guidelines[@hei]. The NCI distributes SAS macros for calculating individual HEI scores. **hei** is a package that implements the HEI scoring algorithms in R[@R], allowing users without SAS licenses to perform these calculations for individuals across several years of National Health and Nutrition Examination Survey (NHANES) studies. The package has been unit tested against HEI scores computed using the SAS functions to validate its performance. Internally, **hei** depends on the **nhanesA** package[@nhanesA] to retrieve necessary data for scoring. HEI scores can be widely applied in the fields of epidemiology, public health, nutrition and beyond[@social_support][@tinnitus][@age], particularly when combined with additional publicly available demographic and behavioral NHANES datasets. 
 
 # References
