### 2017-09-25 vpnagraj/hei#23 @vpnagraj
no publications (yet!) ... but we've included references to examples of publications that use HEI scoring
