### 2017-09-25 vpnagraj/hei#14 @vpnagraj
@maelle that's a good point ... i've included as part of the paper revisions in 4769b93
