### 2017-09-25 vpnagraj/hei#19 @vpnagraj
@maelle thanks, the revised paper in 4769b93 is a little more explicit about the target audience
