### 2017-09-25 vpnagraj/hei#16 @vpnagraj
@maelle a `pkgdown` site would be awesome ... i'll add it as an enhancement for our project
