### 2017-09-26 #417 @maelle
I'm happy with the changes that @timfolsom's coauthor @vpnagraj made. Nearly all issues I've opened are closed, now everything in the checklist is checked. ✅
