### 2017-09-26 #417 @timfolsom
Thank you for the review and the feedback. @vpnagraj and I have incorporated almost all of your suggestions. If it is all the same to you, we would prefer to leave the pkgdown issue open for now.
