### 2017-09-26 #417 @leeper
Thanks, all. This looks good to me. @arfon - over to you.
