### 2017-10-04 b903180
Update paper.md
