### 2017-10-04 #417 @arfon
@vpnagraj - please merge this pull request https://github.com/timfolsom/hei/pull/24
and then could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
