### 2017-10-04 #417 @vpnagraj
@arfon thanks for following up ... i've merged the pull request and published an archive of the package in figshare with the following following DOI:

10.6084/m9.figshare.5469208
