### 2017-10-04 #417 @arfon
@maelle - many thanks for your review and to @leeper for editing this one ✨

@timfolsom @vpnagraj - your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00417 ⚡️ 🚀 💥
